+++ crates/tauri-cli/src/add.rs
@@ -130,6 +130,10 @@ pub fn run(options: Options) -> Result<()> {
     "Builder::new(|pass| todo!()).build()"
   } else if plugin == "localhost" {
     "Builder::new(todo!()).build()"
+  } else if plugin == "single-instance" {
+    "init(|app, args, cwd| {})"
+  } else if plugin == "log" {
+    "Builder::new().level(tauri_plugin_log::log::LevelFilter::Info).build()"
   } else if metadata.builder {
     "Builder::new().build()"
   } else {