+++ crates/tauri-utils/src/platform.rs
@@ -433,7 +433,7 @@ mod tests {
     #[cfg(target_os = "macos")]
     assert!(resource_dir.is_err());
     #[cfg(target_os = "linux")]
-    assert_eq!(resource_dir.unwrap(), PathBuf::from("/usr/lib/my-app"));
+    assert_eq!(resource_dir.unwrap(), PathBuf::from("/usr/lib/MyApp"));
     #[cfg(windows)]
     assert_eq!(resource_dir.unwrap(), path.parent().unwrap());
   }