+++ crates/tauri/src/app.rs
@@ -526,7 +526,7 @@ impl<R: Runtime> AppHandle<R> {
   ///     Ok(())
   ///   });
   /// ```
-  pub fn remove_plugin(&self, plugin: &'static str) -> bool {
+  pub fn remove_plugin(&self, plugin: &str) -> bool {
     self.manager().plugins.lock().unwrap().unregister(plugin)
   }
 
@@ -622,6 +622,17 @@ impl<R: Runtime> AppHandle<R> {
       .runtime_handle
       .set_dock_visibility(visible)
       .map_err(Into::into)
+  }
+
+  /// Change the device event filter mode.
+  ///
+  /// See [App::set_device_event_filter] for details.
+  ///
+  /// ## Platform-specific
+  ///
+  /// See [App::set_device_event_filter] for details.
+  pub fn set_device_event_filter(&self, filter: DeviceEventFilter) {
+    self.runtime_handle.set_device_event_filter(filter);
   }
 }
 
@@ -1621,7 +1632,7 @@ impl<R: Runtime> Builder<R> {
 use tauri::Manager;
 tauri::Builder::default()
   .setup(|app| {
-    let main_window = app.get_window("main").unwrap();
+    let main_window = app.get_webview_window("main").unwrap();
     main_window.set_title("Tauri!")?;
     Ok(())
   });
@@ -2154,6 +2165,26 @@ tauri::Builder::default()
       },
     };
 
+    // The env var must be set before the Runtime is created so that GetAvailableBrowserVersionString picks it up.
+    #[cfg(windows)]
+    {
+      if let crate::utils::config::WebviewInstallMode::FixedRuntime { path } =
+        &manager.config.bundle.windows.webview_install_mode
+      {
+        if let Some(exe_dir) = crate::utils::platform::current_exe()
+          .ok()
+          .and_then(|p| p.parent().map(|p| p.to_path_buf()))
+        {
+          std::env::set_var("WEBVIEW2_BROWSER_EXECUTABLE_FOLDER", exe_dir.join(path));
+        } else {
+          #[cfg(debug_assertions)]
+          eprintln!(
+            "failed to resolve resource directory; fallback to the installed Webview2 runtime."
+          );
+        }
+      }
+    }
+
     #[cfg(any(windows, target_os = "linux"))]
     let mut runtime = if self.runtime_any_thread {
       R::new_any_thread(runtime_args)?
@@ -2230,25 +2261,6 @@ tauri::Builder::default()
 
     app.manage(ChannelDataIpcQueue::default());
     app.handle.plugin(crate::ipc::channel::plugin())?;
-
-    #[cfg(windows)]
-    {
-      if let crate::utils::config::WebviewInstallMode::FixedRuntime { path } =
-        &app.manager.config().bundle.windows.webview_install_mode
-      {
-        if let Ok(resource_dir) = app.path().resource_dir() {
-          std::env::set_var(
-            "WEBVIEW2_BROWSER_EXECUTABLE_FOLDER",
-            resource_dir.join(path),
-          );
-        } else {
-          #[cfg(debug_assertions)]
-          eprintln!(
-            "failed to resolve resource directory; fallback to the installed Webview2 runtime."
-          );
-        }
-      }
-    }
 
     let handle = app.handle();