+++ examples/api/src-tauri/src/tray.rs
@@ -117,6 +117,7 @@ pub fn create_tray<R: Runtime>(app: &tauri::AppHandle<R>) -> tauri::Result<()> {
       {
         let app = tray.app_handle();
         if let Some(window) = app.get_webview_window("main") {
+          let _ = window.unminimize();
           let _ = window.show();
           let _ = window.set_focus();
         }