+++ packages/api/src/window.ts
@@ -1406,6 +1406,22 @@ class Window {
   }
 
   /**
+   * On macOS, Toggles a fullscreen mode that doesn’t require a new macOS space. Returns a boolean indicating whether the transition was successful (this won’t work if the window was already in the native fullscreen).
+   * This is how fullscreen used to work on macOS in versions before Lion. And allows the user to have a fullscreen window without using another space or taking control over the entire monitor.
+   *
+   * On other platforms, this is the same as {@link Window.setFullscreen}.
+   *
+   * @param fullscreen Whether the window should go to simple fullscreen or not.
+   * @returns A promise indicating the success or failure of the operation.
+   */
+  async setSimpleFullscreen(fullscreen: boolean): Promise<void> {
+    return invoke('plugin:window|set_simple_fullscreen', {
+      label: this.label,
+      value: fullscreen
+    })
+  }
+
+  /**
    * Bring the window to front and focus.
    * @example
    * ```typescript
@@ -1418,6 +1434,30 @@ class Window {
   async setFocus(): Promise<void> {
     return invoke('plugin:window|set_focus', {
       label: this.label
+    })
+  }
+
+  /**
+   * Sets whether the window can be focused.
+   *
+   * #### Platform-specific
+   *
+   * - **macOS**: If the window is already focused, it is not possible to unfocus it after calling `set_focusable(false)`.
+   *   In this case, you might consider calling {@link Window.setFocus} but it will move the window to the back i.e. at the bottom in terms of z-order.
+   *
+   * @example
+   * ```typescript
+   * import { getCurrentWindow } from '@tauri-apps/api/window';
+   * await getCurrentWindow().setFocusable(true);
+   * ```
+   *
+   * @param focusable Whether the window can be focused.
+   * @returns A promise indicating the success or failure of the operation.
+   */
+  async setFocusable(focusable: boolean): Promise<void> {
+    return invoke('plugin:window|set_focusable', {
+      label: this.label,
+      value: focusable
     })
   }
 
@@ -2281,6 +2321,8 @@ interface WindowOptions {
   fullscreen?: boolean
   /** Whether the window will be initially focused or not. */
   focus?: boolean
+  /** Whether the window can be focused or not. */
+  focusable?: boolean
   /**
    * Whether the window is transparent or not.
    * Note that on `macOS` this requires the `macos-private-api` feature flag, enabled under `tauri.conf.json > app > macOSPrivateApi`.