+++ crates/tauri-codegen/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## \[2.4.0]
+
+### Dependencies
+
+- Upgraded to `tauri-utils@2.7.0`.
+
 ## \[2.3.1]
 
 ### Dependencies